fix: mdx render error
When a project page was to be rendered it would return an invalid hook
error. This was because the project didn't use mdx/rsc which means it
was still working for the pages directory under the hood. This is
changed to the mdx/rsc.

Also changed the compiling from the serialization to gray-matter to
return only pure markdown